Remove Mixin AP support.
