+++ src/test/groovy/net/fabricmc/loom/test/integration/SimpleProjectTest.groovy
@@ -37,7 +37,6 @@ import net.fabricmc.loom.test.util.GradleProjectTestTrait
 import net.fabricmc.loom.test.util.ServerRunner
 import net.fabricmc.loom.util.ZipUtils
 
-import static net.fabricmc.loom.test.LoomTestConstants.PRE_RELEASE_GRADLE
 import static net.fabricmc.loom.test.LoomTestConstants.STANDARD_TEST_VERSIONS
 import static org.gradle.testkit.runner.TaskOutcome.SUCCESS
 
@@ -84,25 +83,6 @@ class SimpleProjectTest extends Specification implements GradleProjectTestTrait 
 		'ideaSyncTask' 		| _
 		'genEclipseRuns'	| _
 		'vscode'			| _
-	}
-
-	@Unroll
-	def "remap mixins with mixin AP"() {
-		setup:
-		def gradle = gradleProject(project: "simple", version: PRE_RELEASE_GRADLE)
-		gradle.buildGradle << """
-				allprojects {
-					loom.mixin.useLegacyMixinAp = true
-				}
-				""".stripIndent()
-
-		when:
-		def result = gradle.run(task: "build")
-
-		then:
-		result.task(":build").outcome == SUCCESS
-		!result.output.contains("[WARN]  [MIXIN]") // Assert that tiny remapper didnt not have any warnings when remapping
-		gradle.getOutputZipEntry("fabric-example-mod-1.0.0.jar", "META-INF/MANIFEST.MF").contains("Fabric-Loom-Version: 0.0.0+unknown")
 	}
 
 	// Tests that deleted files don't remain in built jars after a rebuild.
