--- src/test/resources/projects/mixinApAutoRefmap/gradle.properties
@@ -1,16 +1,0 @@
-# Done to increase the memory available to gradle.
-org.gradle.jvmargs=-Xmx1G
-
-# Fabric Properties
-	# check these on https://fabricmc.net/versions.html
-	minecraft_version=1.17
-	yarn_mappings=1.17+build.13
-	loader_version=0.11.6
-
-# Mod Properties
-	mod_version = 1.0.0
-	maven_group = com.example
-	archives_base_name = fabric-example-mod
-
-# Dependencies
-	fabric_version=0.36.0+1.17
