--- src/test/resources/projects/mixinApAutoRefmap/src/main/resources/subfolder/subfolder.mixins.json
@@ -1,14 +1,0 @@
-{
-  "required": true,
-  "minVersion": "0.8",
-  "package": "net.fabricmc.example.mixin",
-  "compatibilityLevel": "JAVA_16",
-  "mixins": [
-  ],
-  "client": [
-      "ExampleMixinSubfolder"
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
