--- src/test/resources/projects/mixinApAutoRefmap/src/mixin/resources/blabla.json
@@ -1,14 +1,0 @@
-{
-  "required": true,
-  "minVersion": "0.8",
-  "package": "net.fabricmc.example.mixin",
-  "compatibilityLevel": "JAVA_16",
-  "mixins": [
-  ],
-  "client": [
-      "ExampleMixin0"
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
