--- src/test/resources/projects/mixinApAutoRefmap/src/mixin1/resources/m1_2.mixins.json
@@ -1,14 +1,0 @@
-{
-  "required": true,
-  "minVersion": "0.8",
-  "package": "net.fabricmc.example.mixin",
-  "compatibilityLevel": "JAVA_16",
-  "mixins": [
-  ],
-  "client": [
-      "ExampleMixin1_2"
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
