--- src/test/resources/projects/mixinApSimple/src/main/java/net/fabricmc/example/mixin/ExampleMixin1.java
@@ -1,15 +1,0 @@
-package net.fabricmc.example.mixin;
-
-import net.minecraft.client.gui.screen.ChatScreen;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-
-@Mixin(ChatScreen.class)
-public class ExampleMixin1 {
-	@Inject(at = @At("HEAD"), method = "init()V")
-	private void init(CallbackInfo info) {
-		System.out.println("This line is printed by an example mod mixin!");
-	}
-}
